+++ react/utils/constants.ts
@@ -1,6 +1,7 @@
 export const PAGE_SIZE = 15
 export const INITIAL_PAGE = 1
 export const MAX_PAGE_SIZE = 100
+export const MIN_TABLE_SIZE = 5
 
 // Our component treats 0 as infinity
 export const COMMISSIONS_TABLE_EXPORT_LIMIT = 0