+++ CHANGELOG.md
@@ -7,12 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Minimum table size for the profile
+- Table empty state
+- Filter by orderId, status and date for the profile table
+- `notifyOnNetworkStatusChange` prop to the affiliateOrders query
+
 ## [0.55.0] - 2022-07-22
+
 ### Added
 
 - Created and added the `Tooltip` component to be used inside `Totalizer`
 
 ## [0.54.0] - 2022-07-18
+
 ### Added
 
 - Created and added the `ProfileValidator` component and `profile-validator` block